Implement shared components for login and registration pages
